+++ README.md
@@ -1,1 +1,11 @@
-MicroProcessor-LPCXpresso-1769
+# Microprocessor and interfacing course - IC KMITL 2018
+Code for LPC1769 LPCXPRESSO microprocessor labratory
+
+![lpcxpressoBoard]()
+
+![pin]()
+
+
+
+
+
