+++ lab4/Debug/src/crp.d
@@ -1,0 +1,1 @@
+src/crp.o src/crp.d: ../src/crp.c
