+++ lab5/Debug/objects.mk
@@ -1,0 +1,8 @@
+################################################################################
+# Automatically-generated file. Do not edit!
+################################################################################
+
+USER_OBJS :=
+
+LIBS := -llpc_board_nxp_lpcxpresso_1769 -llpc_chip_175x_6x
+
