+++ lab5/src/timer.c
@@ -1,0 +1,101 @@
+/*
+ * @brief Blinky example using timers
+ *
+ * @note
+ * Copyright(C) NXP Semiconductors, 2014
+ * All rights reserved.
+ *
+ * @par
+ * Software that is described herein is for illustrative purposes only
+ * which provides customers with programming information regarding the
+ * LPC products.  This software is supplied "AS IS" without any warranties of
+ * any kind, and NXP Semiconductors and its licensor disclaim any and
+ * all warranties, express or implied, including all implied warranties of
+ * merchantability, fitness for a particular purpose and non-infringement of
+ * intellectual property rights.  NXP Semiconductors assumes no responsibility
+ * or liability for the use of the software, conveys no license or rights under any
+ * patent, copyright, mask work right, or any other intellectual property rights in
+ * or to any products. NXP Semiconductors reserves the right to make changes
+ * in the software without notification. NXP Semiconductors also makes no
+ * representation or warranty that such application will be suitable for the
+ * specified use without further testing or modification.
+ *
+ * @par
+ * Permission to use, copy, modify, and distribute this software and its
+ * documentation is hereby granted, under NXP Semiconductors' and its
+ * licensor's relevant copyrights in the software, without fee, provided that it
+ * is used in conjunction with NXP Semiconductors microcontrollers.  This
+ * copyright, permission, and disclaimer notice must appear in all copies of
+ * this code.
+ */
+
+#include "board.h"
+#include <stdio.h>
+
+/*****************************************************************************
+ * Private types/enumerations/variables
+ ****************************************************************************/
+
+#define TICKRATE_HZ1 (5)	/* 11 ticks per second */
+
+/*****************************************************************************
+ * Public types/enumerations/variables
+ ****************************************************************************/
+
+/*****************************************************************************
+ * Private functions
+ ****************************************************************************/
+
+/*****************************************************************************
+ * Public functions
+ ****************************************************************************/
+
+/**
+ * @brief	Handle interrupt from 32-bit timer
+ * @return	Nothing
+ */
+void TIMER0_IRQHandler(void)
+{
+	if (Chip_TIMER_MatchPending(LPC_TIMER0, 1)) {
+		Chip_TIMER_ClearMatch(LPC_TIMER0, 1);
+		Board_LED_Toggle(0);
+	}
+	printf("test");
+}
+
+/**
+ * @brief	main routine for blinky example
+ * @return	Function should not exit.
+ */
+int main(void)
+{
+	uint32_t timerFreq;
+
+	/* Generic Initialization */
+	SystemCoreClockUpdate();
+	Board_Init();
+
+	/* Enable timer 1 clock */
+	Chip_TIMER_Init(LPC_TIMER0);
+
+	/* Timer rate is system clock rate */
+	timerFreq = Chip_Clock_GetSystemClockRate();
+
+	/* Timer setup for match and interrupt at TICKRATE_HZ */
+	Chip_TIMER_Reset(LPC_TIMER0);
+	Chip_TIMER_MatchEnableInt(LPC_TIMER0, 1);
+	Chip_TIMER_SetMatch(LPC_TIMER0, 1, timerFreq / (8*5));
+	Chip_TIMER_ResetOnMatchEnable(LPC_TIMER0, 1);
+	Chip_TIMER_Enable(LPC_TIMER0);
+
+	/* Enable timer interrupt */
+	NVIC_ClearPendingIRQ(TIMER0_IRQn);
+	NVIC_EnableIRQ(TIMER0_IRQn);
+
+	/* LEDs toggle in interrupt handlers */
+	while (1) {
+		__WFI();
+	}
+
+	return 0;
+}
