--- lab_4/Debug/lab_4_Debug_library.ld
@@ -1,14 +1,0 @@
-/*
- * GENERATED FILE - DO NOT EDIT
- * (c) Code Red Technologies Ltd, 2008-2013
- * (c) NXP Semiconductors 2013-2018
- * Generated linker script file for LPC1769
- * Created from library.ldt by FMCreateLinkLibraries
- * Using Freemarker v2.3.23
- * LPCXpresso v8.2.2 [Build 650] [2016-09-09]  on Sep 7, 2018 10:41:14 PM
- */
-
-GROUP (
-  libcr_c.a
-  libcr_eabihelpers.a
-)
