--- lab_4/Debug/src/cr_startup_lpc175x_6x.d
@@ -1,2 +1,0 @@
-src/cr_startup_lpc175x_6x.o src/cr_startup_lpc175x_6x.d: \
- ../src/cr_startup_lpc175x_6x.c
