--- lab_4/Debug/src/crp.d
@@ -1,1 +1,0 @@
-src/crp.o src/crp.d: ../src/crp.c
